refactor: Rough refactor of [id] including new files creation and useServer hook creation
